nice mermao

+++ TCC-LOSPACO/Models/Package.cs
@@ -2,7 +2,7 @@
 
 namespace TCC_LOSPACO.Models {
     public class Package {
-        public uint Id { get; private set; }
+        public ushort Id { get; private set; }
         public string Name { get; private set; }
         public string MinifiedDesc { get; private set; }
         public string Description { get; private set; }
@@ -10,7 +10,7 @@ namespace TCC_LOSPACO.Models {
         public decimal Price { get; private set; }
         public List<Service> Services { get; private set; }
 
-        public Package(uint id, string name, string minifiedDesc, string description, byte[] image, decimal price, List<Service> services) {
+        public Package(ushort id, string name, string minifiedDesc, string description, byte[] image, decimal price, List<Service> services) {
             Id = id;
             Name = name;
             MinifiedDesc = minifiedDesc;
